Use long variables for the value of IConst

+++ dartagnan/src/main/java/com/dat3m/dartagnan/expression/BExpr.java
@@ -13,7 +13,7 @@ public abstract class BExpr implements ExprInterface {
     }
 
     @Override
-    public int getIntValue(Event e, Model model, Context ctx){
+    public long getIntValue(Event e, Model model, Context ctx){
         return getBoolValue(e, model, ctx) ? 1 : 0;
     }
