+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/Register.java
@@ -88,8 +88,8 @@ public class Register extends IExpr implements ExprInterface {
 	}
 
 	@Override
-	public int getIntValue(Event e, Model model, Context ctx){
-		return Integer.parseInt(model.getConstInterp(toZ3Int(e, ctx)).toString());
+	public long getIntValue(Event e, Model model, Context ctx){
+		return Long.parseLong(model.getConstInterp(toZ3Int(e, ctx)).toString());
 	}
 
 	@Override
